fix(privacy-export): move InvalidExportEntryPathException to Exceptions/ folder
Archi test FileOrganizationTests.Exception_classes_should_reside_in_Exceptions_folder
required the exception under Exceptions/. Move it under Granit.Privacy/Exceptions/
(module-root convention used by all other modules) and update the sanitizer to
reference the new namespace.
