+++ src/Granit.Privacy/Exceptions/InvalidExportEntryPathException.cs
@@ -1,9 +1,11 @@
-namespace Granit.Privacy.DataExport.Sanitization;
+using Granit.Privacy.DataExport.Sanitization;
+
+namespace Granit.Privacy.Exceptions;
 
 /// <summary>
 /// Thrown by <see cref="EntryPathSanitizer.Sanitize"/> when an
-/// <see cref="Fragments.ExportFragment.EntryPath"/> fails validation. Closes <c>VULN-004</c>
-/// (zip-slip via attacker-controlled path) by failing fast at the provider boundary.
+/// <see cref="DataExport.Fragments.ExportFragment.EntryPath"/> fails validation. Guards against
+/// zip-slip via attacker-controlled paths by failing fast at the provider boundary.
 /// </summary>
 public sealed class InvalidExportEntryPathException(string entryPath, string reason)
     : Exception($"Invalid export entry path '{entryPath}': {reason}")
